feat: creating orders in database

+++ models/cart.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const path = require('path');
-const db = require('../util/database')
+const db = require('../util/database');
+const { log } = require('console');
 
 const p = path.join(
   path.dirname(process.mainModule.filename),
@@ -13,15 +14,41 @@ const p = path.join(
 module.exports = class Cart {
 
   static deleteProductFromCart(id) {
-    console.log(id);
-    return db.none('DELETE FROM cart WHERE product_id = $1', id)
+    return db.none('DELETE FROM cart WHERE product_id = $1 AND user_id = 1', id)
   }
 
-  static getCart(cb) {
-    return db.any('SELECT p.product_id, title, imgurl, price, description, c.qty FROM products p INNER JOIN cart c ON c.product_id = p.product_id ORDER BY qty');
+  static getCart() {
+    return db.any('SELECT p.product_id, title, imgurl, price, description, c.qty FROM products p INNER JOIN cart c ON c.product_id = p.product_id WHERE c.user_id = 1 ORDER BY qty');
   }
 
   static AddProductToCart(id) {
-    return db.none('INSERT INTO cart (product_id, qty) VALUES($1, 1) ON CONFLICT(product_id) DO UPDATE SET qty = 2', id);
+    const promise = db.any('SELECT product_id FROM cart WHERE product_id = $1 AND user_id = 1', id).then(res => {
+      if (res.length > 0) {
+        return db.none('UPDATE cart SET qty = qty + 1 WHERE product_id = $1 AND user_id = 1', id)
+      } else {
+        return db.none('INSERT INTO cart (user_id, product_id, qty) VALUES(1, $1, 1)', id)
+      }
+    }).catch(err => {
+      console.log(err);
+    }) 
+    return Promise.all([promise])
+  }
+
+  static emptyCart() {
+    return db.none('DELETE FROM cart WHERE user_id = 1');
+  }
+
+  static createOrder(totalPrice) {
+    const createOrderPromise = db.one('INSERT INTO orders (user_id, total) VALUES(1, $1) RETURNING order_id', totalPrice);
+    createOrderPromise.then((res) => {
+      let orderId = res.order_id;
+      console.log(orderId, 'order id');
+    });
+    const getProductsIdPromise = db.any('SELECT p.product_id FROM products p INNER JOIN cart c ON c.product_id = p.product_id WHERE c.user_id = 1 ORDER BY c.cart_id');
+    getProductsIdPromise.then((res) => {
+      console.log(res); 
+    })
+
+    Promise.all([createOrderPromise, getProductsIdPromise])
   }
 };
